设备管理删除确认，提示显示

+++ src/components/equipments/verification_plan.vue
@@ -330,29 +330,35 @@
 					});
 					return;
 				} else {
-					var changeUser = selData[0];
-					var id = changeUser.ID;
-					var url = this.basic_url + '/api-apps/app/checkPlan/' + id;
-					this.$axios.delete(url, {}).then((res) => {//.delete 传数据方法
-						//resp_code == 0是后台返回的请求成功的信息
-						if(res.data.resp_code == 0) {
+					this.$confirm('确定要删除此数据吗?', '提示', {
+						confirmButtonText: '确定',
+						cancelButtonText: '取消',
+						type: 'warning'
+					}).then(() => {
+						var changeUser = selData[0];
+						var id = changeUser.ID;
+						var url = this.basic_url + '/api-apps/app/checkPlan/' + id;
+						this.$axios.delete(url, {}).then((res) => {//.delete 传数据方法
+							//resp_code == 0是后台返回的请求成功的信息
+							if(res.data.resp_code == 0) {
+								this.$message({
+									message: '删除成功',
+									type: 'success'
+								});
+								this.requestData();
+							}else{
+								this.$message({
+									message: res.data.resp_msg,
+									type: 'success'
+								});
+							}
+						}).catch((err) => {
 							this.$message({
-								message: '删除成功',
-								type: 'success'
+								message: '网络错误，请重试',
+								type: 'error'
 							});
-							this.requestData();
-						}else{
-							this.$message({
-								message: res.data.resp_msg,
-								type: 'success'
-							});
-						}
-					}).catch((err) => {
-						this.$message({
-							message: '网络错误，请重试',
-							type: 'error'
 						});
-					});
+					}).catch(() => {});   
 				}
 			},
 			// 导入
